// Qanat.Web/src/app/shared/generated/model/parcel-water-account-history-simple-dto.ts
export class ParcelWaterAccountHistorySimpleDto { 
    ParcelWaterAccountHistoryID?: number;
    GeographyID?: number;
    ParcelID?: number;
    ReportingPeriodID?: number;
    FromWaterAccountID?: number | null;
    FromWaterAccountNumber?: number | null;
    FromWaterAccountName?: string | null;
    ToWaterAccountID?: number | null;
    ToWaterAccountNumber?: number | null;
    ToWaterAccountName?: string | null;
    Reason?: string | null;
    CreateUserID?: number;
    CreateDate?: string;
    ParcelNumber?: string | null;
    ReportingPeriodName?: string | null;
    CreateUserFullName?: string | null;
    readonly FromWaterAccountNumberAndName?: string | null;
    readonly ToWaterAccountNumberAndName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}