// Qanat.Web/src/app/shared/generated/model/usage-location-parcel-history-dto.ts
export class UsageLocationParcelHistoryDto { 
    UsageLocationParcelHistoryID?: number;
    GeographyID?: number;
    UsageLocationID?: number;
    UsageLocationName?: string | null;
    ReportingPeriodID?: number;
    ReportingPeriodName?: string | null;
    FromParcelID?: number | null;
    FromParcelNumber?: string | null;
    ToParcelID?: number | null;
    ToParcelNumber?: string | null;
    CreateDate?: string;
    CreateUserFullName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}